Fix final exponentiation for multiple pairings
The arguments in `conditionally_select` were inverted due to the fact
the order is inverted compared to gnark's code. I've added a test that
performs 2 pairings. Due to the large circuit size, it takes a long time
to complete for >2 pairings, but it seems to be working with 3 pairings
too. The multiple pairings test is also commented out for perf reasons.

Also added a `reduce` call before a `conditionally_select` to ensure
that the limb counts between the `dummy` variable and `e.c1` are the
same.